plugins/hello-world/skills/orchestrate/requirements.txt
rich>=13.0
typer>=0.9
pydantic>=2.0
pyyaml>=6.0
aiofiles>=23.0
